+++ src/app/not-found.tsx
@@ -1,0 +1,11 @@
+import Layout from '@/components/Layout';
+
+export default function NotFound() {
+  return (
+    <Layout>
+      <main className="px-5 bg-blue-50 pt-5 pb-20">
+        <h1>ページが見つかりません</h1>
+      </main>
+    </Layout>
+  );
+}
