--- src/pages/_app.tsx
@@ -1,9 +1,0 @@
-// TODO: AppRouter 移行後削除
-
-import '../styles/globals.css';
-
-const MyApp = ({ Component, pageProps }) => {
-  return <Component {...pageProps} />;
-};
-
-export default MyApp;
